Major bug fix for non-functional updates on pre-visited rooms.
I feel the need to rant about this one. Apparently sails is very
particular about context to which a socket is subscribed. In other
words, if a socket is unsubscribed from everything, it can never be
re-subscribed to any context no matter how hard you try. It is therefore
important that from now on we make sure to ALWAYS subscribe to specific
object contexts, and make sure to use those same (or a subset of those)
contexts when unsubscribing.

Other fixes include minor tidying up and conversion of some controller
methods to avoid infinite callback syndrome.