Added configuration for a local mongo db. It surprisingly worked.
It worked the first time too!

For future reference, here are instructions on how to start the mongo
server:

    mongod --dbpath [the path to where the databases are located]

You will need a database named 'betrayal' at that location.

To verify you can connect to that database, use the mongo shell.

    mongo
    use betrayal

To start the sails app to connect to this database, set your NODE_ENV
variable to be 'local_mongo'.

    export NODE_ENV='local_mongo'